Use an HOC to wrap the various react-router component
