+++ src/modules/app/components/inner-nav/base-inner-nav-pure.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'modules/common/containers/sticky-params-link'
+import { Link } from 'modules/common/containers/sticky-params-components'
 import classNames from 'classnames'
 import { mobileMenuStates } from 'modules/app/components/app/app'
