+++ src/modules/app/components/inner-nav/base-inner-nav.jsx
@@ -1,7 +1,7 @@
 /* eslint class-methods-use-this: 0 */ // need "blank" class methods to exist to avoid potential crash
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'modules/common/containers/sticky-params-link'
+import { Link } from 'modules/common/containers/sticky-params-components'
 import classNames from 'classnames'
 import { mobileMenuStates } from 'modules/app/components/app/app'
