+++ src/modules/categories/components/categories-view/categories-view.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'modules/common/containers/sticky-params-link'
+import { Link } from 'modules/common/containers/sticky-params-components'
 import { Helmet } from 'react-helmet'
 import { createBigNumber } from 'utils/create-big-number'
 import CategoryList from 'modules/categories/components/category-list/category-list'
