+++ src/modules/common/components/main-error-boundary/main-error-boundary.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { Link } from 'modules/common/containers/sticky-params-link'
+import { Link } from 'modules/common/containers/sticky-params-components'
 import PropTypes from 'prop-types'
 import makePath from 'modules/routes/helpers/make-path'
 import { DEFAULT_VIEW } from 'modules/routes/constants/views'
