+++ src/modules/common/components/paginator/paginator.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'modules/common/containers/sticky-params-link'
+import { Link } from 'modules/common/containers/sticky-params-components'
 
 import { PAGINATION_PARAM_NAME } from 'modules/routes/constants/param-names'
